frontend/requirements.txt
flask
elasticsearch
numpy
chromadb
openai
python-dotenv
gunicorn
